Refactor barcode, QR code, metadata, and short URL APIs: move business logic to dedicated service files for better organization and maintainability.
