+++ app/api/metadata.py
@@ -1,6 +1,6 @@
 from fastapi import APIRouter, HTTPException, Query
 from pydantic import AnyUrl
-from app.utils.metadata_utils import fetch_url_metadata
+from app.services.metadata_service import fetch_metadata_logic
 
 router = APIRouter(
     prefix="/metadata",
@@ -10,7 +10,6 @@ router = APIRouter(
 @router.get("/", status_code=200)
 async def get_metadata(url: AnyUrl = Query(..., description="Target URL to fetch metadata")):
     try:
-        meta = fetch_url_metadata(str(url))
-        return meta
+        return fetch_metadata_logic(str(url))
     except Exception as e:
         raise HTTPException(status_code=500, detail=f"Failed to fetch metadata: {str(e)}")
